Ravenous: Implement private CORS server
Need to create a private server to replace the previously used public one as it was being discontinued. Server attaches the correct headers to a API request to make it CORS compliant. See separate GitHib repo.
Issues with hosting the server that on Namecheap. Instead, created a Heroku account and launched the server there.