tileSelector: Ravenous link opens Heroku app
Rather than open a subdirectory, link for the Ravenous app should open the herokuapp subdomain where the app is hosted.
It wasn't possible to connect the /ravenous subdirectory to the herokuapp subdomain. I would have to create a subdomain for the website to connect it up, but that would not be much better than this solution.