Gitignore: Ignore .runtimeconfig.json
The .runtimeconfig.json file contains the password for the email router and is referenced by the environment variables when testing locally.
The file is generated by downloading the website keys from the Service Accounts pane of the Google Cloud Console and then running the script from the commit following this one.